Rename Card components
