+++ src/templates/components/CapiCard.svelte
