+++ src/templates/components/ManualCard.svelte
