+++ src/templates/csr/capi-single-paidfor/index.svelte
@@ -9,7 +9,7 @@
 	import type { Single } from '$lib/types/capi';
 	import type { GAMVariable } from '$lib/gam';
 
-	import Card from '$templates/components/Card.svelte';
+	import CapiCard from '$templates/components/CapiCard.svelte';
 	import PaidForHeader from '$templates/components/PaidForHeader.svelte';
 	import { addTrackingPixel, isValidReplacedVariable } from '$lib/gam';
 	import Resizer from '$templates/components/Resizer.svelte';
@@ -36,7 +36,7 @@
 			{ComponentTitle}
 			{SeriesUrl}
 		/>
-		<Card {single} />
+		<CapiCard {single} />
 	</aside>
 	<Resizer {height} />
 {:catch}
